@@ -5,9 +5,10 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "build": "set NODE_ENV=production&& webpack",
+    "build": "yarn clean && set NODE_ENV=production&& webpack",
     "watch": "yarn clean && webpack --watch --mode=development",
-    "clean": "rimraf ./dist/healthcheck*",
+    "devbuild": "yarn clean && webpack --mode=development",
+    "clean": "rimraf ./dist/*.js && rimraf ./dist/*.ttf",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
@@ -19,34 +20,38 @@
     "@lazy-copilot/datetimepicker": "^0.1.47",
     "babel-polyfill": "^6.26.0",
     "chart.js": "^2.8.0",
-    "style-loader": "^0.23.1",
+    "dagre": "0.8.5",
+    "es6-promise": "^4.2.8",
+    "jointjs": "3.1.1",
+    "monaco-editor": "^0.20.0",
     "vue": "^2.6.9",
     "vue-class-component": "^7.0.1",
-    "vue-monaco-cdn": "^1.0.1",
     "vue-property-decorator": "^8.0.0",
-    "vue-router": "^3.0.2",
+    "vue-router": "^3.2.0",
     "vuedraggable": "^2.23.0",
     "vuetify": "^1.5.6",
-    "xregexp": "4.2.4",
-    "jointjs": "3.1.1",
-    "dagre": "0.8.5"
+    "vuex": "^3.4.0",
+    "xregexp": "4.2.4"
   },
   "devDependencies": {
     "@types/chart.js": "2.8.7",
-    "@types/xregexp": "3.0.30",
     "@types/dagre": "0.7.42",
+    "@types/xregexp": "3.0.30",
     "css-loader": "^2.1.1",
-    "ts-loader": "^5.3.3",
+    "style-loader": "^0.23.1",
+    "file-loader": "^6.0.0",
     "deepmerge": "^4.2.1",
     "fibers": "^4.0.2",
+    "monaco-editor-webpack-plugin": "^1.9.0",
+    "rimraf": "^3.0.0",
     "sass": "^1.23.1",
     "sass-loader": "^8.0.0",
-    "typescript": "^3.3.3333",
+    "ts-loader": "^5.3.3",
+    "typescript": "^3.7.5",
     "vue-loader": "^15.7.0",
     "vue-template-compiler": "^2.6.9",
     "webpack": "^4.29.6",
     "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-cli": "^3.3.0",
-    "rimraf": "^3.0.0"
+    "webpack-cli": "^3.3.0"
   }
 }
