@@ -1,23 +1,11 @@
 <!-- src/components/result_data/data_types/abstractions/CodeComponent.vue -->
 <template>
     <div>
-        <div class="code-editor-loader" v-if="!isEditorInited">
-            <v-progress-circular
-                class="mr-2"
-                indeterminate
-                color="primary"
-            />
-            Loading {{ language }}..
-        </div>
-
-        <monaco-editor
+        <editor-component
             class="editor"
             :class="`${(fullscreen ? 'fullscreen' : '')}`"
             v-model="data.Data"
-            v-on:editorDidMount="onEditorDidMount"
-            :theme="monacoTheme"
-            :options="monacoOptions"
-            :syncInput="true"
+            :read-only="true"
             :language="language"
             ref="editor"/>
     </div>
@@ -26,12 +14,11 @@
 <script lang="ts">
 import { Vue, Component, Prop, Watch } from "vue-property-decorator";
 import TestResultDataDumpViewModel from "../../../../../models/TestSuite/TestResultDataDumpViewModel";
-//@ts-ignore
-import MonacoEditor from 'vue-monaco-cdn'
+import EditorComponent from "../../../../Common/EditorComponent.vue";
 
 @Component({
     components: {
-        MonacoEditor
+        EditorComponent
     }
 })
 export default class CodeEditor extends Vue {
@@ -42,48 +29,24 @@ export default class CodeEditor extends Vue {
     @Prop({ required: true })
     language!: string;
 
-    isEditorInited: boolean = false;
-    editor!: any;
-    monacoTheme: string = "vs-dark";
-    monacoOptions: any = {
-        readOnly: true
-    };
-
     mounted(): void {
-        window.addEventListener('resize', this.onWindowResize);
-        setTimeout(() => {
-            this.refreshSize();
-        }, 10);
-    }
-
-    beforeDestroy(): void {
-        window.removeEventListener('resize', this.onWindowResize)
     }
 
     refreshSize(): void {
-        if (this.editor) {
-            this.editor.layout();
-        }
+        const editor: EditorComponent = <EditorComponent>this.$refs.editor;
+        editor.refreshSize();
     }
     
     ////////////////////////////////////////////////////////////
     //// EVENTHANDLERS /////////////////////////////////////////
     ////////////////////////////////////////////////////////////
-    onEditorDidMount(editor: any): void {
-        this.editor = (<any>this.$refs.editor)["editor"];
-        this.editor.layout();
-        this.editor.updateOptions( { glyphMargin: true });
-        this.isEditorInited = true;
-    }
-
-    onWindowResize(): void {
-        this.refreshSize();
-    }
-
     @Watch("fullscreen")
     onFullscreenChanged(): void {
         this.refreshSize();
         this.$nextTick(() => this.refreshSize());
+        setTimeout(() => {
+            this.refreshSize();
+        }, 100);
     }
 }
 </script>
