@@ -250,20 +250,18 @@ namespace HealthCheck.Core.Util
 
                 lock (_fileLock)
                 {
-                    using (FileStream fileStream = File.Open(FilePath, FileMode.Open, FileAccess.Read, FileShare.ReadWrite))
-                    using (ReverseStreamReader streamReader = new ReverseStreamReader(fileStream))
+                    using FileStream fileStream = File.Open(FilePath, FileMode.Open, FileAccess.Read, FileShare.ReadWrite);
+                    using ReverseStreamReader streamReader = new ReverseStreamReader(fileStream);
+                    string row;
+                    while ((row = streamReader.ReadLine()) != null)
                     {
-                        string row;
-                        while ((row = streamReader.ReadLine()) != null)
-                        {
-                            if (string.IsNullOrWhiteSpace(row))
-                                continue;
+                        if (string.IsNullOrWhiteSpace(row))
+                            continue;
 
-                            var item = DeserializeRow(row);
-                            if (item != null)
-                            {
-                                yield return item;
-                            }
+                        var item = DeserializeRow(row);
+                        if (item != null)
+                        {
+                            yield return item;
                         }
                     }
                 }
@@ -272,21 +270,19 @@ namespace HealthCheck.Core.Util
             {
                 lock (_fileLock)
                 {
-                    using (FileStream fileStream = File.Open(FilePath, FileMode.Open, FileAccess.Read, FileShare.ReadWrite))
-                    using (BufferedStream bufferedStream = new BufferedStream(fileStream))
-                    using (StreamReader streamReader = new StreamReader(bufferedStream))
+                    using FileStream fileStream = File.Open(FilePath, FileMode.Open, FileAccess.Read, FileShare.ReadWrite);
+                    using BufferedStream bufferedStream = new BufferedStream(fileStream);
+                    using StreamReader streamReader = new StreamReader(bufferedStream);
+                    string row;
+                    while ((row = streamReader.ReadLine()) != null)
                     {
-                        string row;
-                        while ((row = streamReader.ReadLine()) != null)
-                        {
-                            if (string.IsNullOrWhiteSpace(row))
-                                continue;
+                        if (string.IsNullOrWhiteSpace(row))
+                            continue;
 
-                            var item = DeserializeRow(row);
-                            if (item != null)
-                            {
-                                yield return item;
-                            }
+                        var item = DeserializeRow(row);
+                        if (item != null)
+                        {
+                            yield return item;
                         }
                     }
                 }
@@ -336,7 +332,7 @@ namespace HealthCheck.Core.Util
             }
 
             // Less than min time since last cleanup => abort
-            if (LastCleanupPerformedAt != null && (DateTime.Now - LastCleanupPerformedAt) < RetentionOptions.MinimumCleanupInterval)
+            if (LastCleanupPerformedAt != null && (DateTime.Now.ToUniversalTime() - LastCleanupPerformedAt?.ToUniversalTime()) < RetentionOptions.MinimumCleanupInterval)
             {
                 return false;
             }
@@ -349,8 +345,8 @@ namespace HealthCheck.Core.Util
             LastCleanupPerformedAt = DateTime.Now;
             if (RetentionOptions.MaxItemAge != null && RetentionOptions.ItemTimestampSelector != null)
             {
-                var threshold = DateTime.Now - RetentionOptions.MaxItemAge;
-                DeleteWhere(x => RetentionOptions.ItemTimestampSelector(x) <= threshold);
+                var threshold = DateTime.Now.ToUniversalTime() - RetentionOptions.MaxItemAge;
+                DeleteWhere(x => RetentionOptions.ItemTimestampSelector(x).ToUniversalTime() <= threshold);
             }
         }
 
