@@ -1,5 +1,5 @@
-﻿using HealthCheck.Core.Attributes;
-using HealthCheck.Core.Entities;
+﻿using HealthCheck.Core.Modules.Tests.Attributes;
+using HealthCheck.Core.Modules.Tests.Models;
 using HealthCheck.WebUI.Serializers;
 using System;
 using System.Linq;
