@@ -1,25 +1,45 @@
 ﻿using HealthCheck.Core.Abstractions;
 using HealthCheck.Core.Attributes;
-using HealthCheck.Core.Entities;
-using HealthCheck.Core.Enums;
 using HealthCheck.Core.Extensions;
+using HealthCheck.Core.Modules.AuditLog;
+using HealthCheck.Core.Modules.AuditLog.Abstractions;
 using HealthCheck.Core.Modules.Dataflow;
-using HealthCheck.Core.Modules.Diagrams.FlowCharts;
-using HealthCheck.Core.Modules.Diagrams.SequenceDiagrams;
+using HealthCheck.Core.Modules.Dataflow.Abstractions;
+using HealthCheck.Core.Modules.Dataflow.Models;
+using HealthCheck.Core.Modules.Dataflow.Services;
+using HealthCheck.Core.Modules.Documentation;
+using HealthCheck.Core.Modules.Documentation.Services;
 using HealthCheck.Core.Modules.EventNotifications;
+using HealthCheck.Core.Modules.EventNotifications.Abstractions;
 using HealthCheck.Core.Modules.EventNotifications.Notifiers;
+using HealthCheck.Core.Modules.EventNotifications.Services;
+using HealthCheck.Core.Modules.LogViewer;
 using HealthCheck.Core.Modules.Settings;
+using HealthCheck.Core.Modules.Settings.Abstractions;
+using HealthCheck.Core.Modules.Settings.Attributes;
+using HealthCheck.Core.Modules.SiteEvents;
+using HealthCheck.Core.Modules.SiteEvents.Abstractions;
+using HealthCheck.Core.Modules.SiteEvents.Enums;
+using HealthCheck.Core.Modules.SiteEvents.Models;
+using HealthCheck.Core.Modules.SiteEvents.Services;
+using HealthCheck.Core.Modules.Tests;
 using HealthCheck.Core.Services;
 using HealthCheck.Core.Services.Models;
 using HealthCheck.Core.Util;
 using HealthCheck.DevTest._TestImplementation;
 using HealthCheck.DevTest._TestImplementation.Dataflow;
 using HealthCheck.DevTest._TestImplementation.EventNotifier;
+using HealthCheck.Module.DynamicCodeExecution.Abstractions;
+using HealthCheck.Module.DynamicCodeExecution.Models;
+using HealthCheck.Module.DynamicCodeExecution.Module;
+using HealthCheck.Module.DynamicCodeExecution.PreProcessors;
+using HealthCheck.Module.DynamicCodeExecution.Storage;
+using HealthCheck.Module.DynamicCodeExecution.Validators;
+using HealthCheck.Modules.DevModule;
 using HealthCheck.RequestLog.Services;
 using HealthCheck.WebUI.Abstractions;
 using HealthCheck.WebUI.Models;
 using HealthCheck.WebUI.Services;
-using HealthCheck.WebUI.ViewModels;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -28,6 +48,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Web;
 using System.Web.Hosting;
+using System.Web.Http;
 using System.Web.Mvc;
 
 namespace HealthCheck.DevTest.Controllers
@@ -49,154 +70,113 @@ namespace HealthCheck.DevTest.Controllers
         private static readonly FlatFileEventSinkKnownEventDefinitionsStorage EventSinkNotificationDefinitionStorage
             = new FlatFileEventSinkKnownEventDefinitionsStorage(@"c:\temp\eventconfig_defs.json");
         private IHealthCheckSettingsService SettingsService { get; set; } = new FlatFileHealthCheckSettingsService<TestSettings>(@"C:\temp\settings.json");
+        private IDataflowService<RuntimeTestAccessRole> DataflowService { get; set; }
+        private IEventDataSink EventSink { get; set; }
         private static bool ForceLogout { get; set; }
 
         #region Init
         public DevController()
-            : base(assemblyContainingTests: typeof(DevController).Assembly) {
-            if (_siteEventService == null)
-            {
+        {
+            InitServices();
 
-                _siteEventService = CreateSiteEventService();
-                _auditEventService = CreateAuditEventService();
-            }
-
-            Services.SiteEventService = _siteEventService;
-            Services.AuditEventService = _auditEventService;
-            Services.LogSearcherService = CreateLogSearcherService();
-            Services.RequestLogService = RequestLogServiceAccessor.Current;
-            Services.SequenceDiagramService = new DefaultSequenceDiagramService(new DefaultSequenceDiagramServiceOptions()
-            {
-                DefaultSourceAssemblies = new[] { typeof(DevController).Assembly }
-            });
-            Services.FlowChartsService = new DefaultFlowChartService(new DefaultFlowChartServiceOptions()
-            {
-                DefaultSourceAssemblies = new[] { typeof(DevController).Assembly }
-            });
-            Services.DataflowService = new DefaultDataflowService<RuntimeTestAccessRole>(new DefaultDataflowServiceOptions<RuntimeTestAccessRole>()
-            {
-                Streams = new IDataflowStream<RuntimeTestAccessRole>[]
+            UseModule(new HCDynamicCodeExecutionModule(new HCDynamicCodeExecutionModuleOptions() {
+                TargetAssembly = typeof(DevController).Assembly,
+                ScriptStorage = new FlatFileDynamicCodeScriptStorage(@"C:\temp\DCE_Scripts.json"),
+                PreProcessors = new IDynamicCodePreProcessor[]
                 {
-                    testStreamA,
-                    testStreamB,
-                    testStreamC,
-                    simpleStream,
-                    memoryStream,
-                    otherStream1,
-                    otherStream2
+                    new BasicAutoCreateUsingsPreProcessor(typeof(DevController).Assembly),
+                    new WrapUsingsInRegionPreProcessor(),
+                    new FuncPreProcessor("'woot' replacer", (p, code) => code.Replace("woot", "w00t"), "Replaces any instances of 'woot' with 'w00t'"),
+                    new FuncPreProcessor("Bad words be gone", (p, code) => code.Replace(" bad ", "***"), canBeDisabled: false)
+                },
+                Validators = new IDynamicCodeValidator[]
+                {
+                    new FuncCodeValidator((code) =>
+                        code.Contains("format c:")
+                            ? DynamicCodeValidationResult.Deny("No format pls")
+                            : DynamicCodeValidationResult.Allow()
+                    )
                 }
-            });
-            Services.SettingsService = SettingsService;
-            Services.EventSink = new DefaultEventDataSink(EventSinkNotificationConfigStorage, EventSinkNotificationDefinitionStorage)
-                .AddNotifier(new WebHookEventNotifier())
-                .AddNotifier(new MyNotifier())
-                .AddNotifier(new SimpleNotifier())
-                .AddPlaceholder("NOW", () => DateTime.Now.ToString())
-                .AddPlaceholder("ServerName", () => Environment.MachineName);
-            (Services.EventSink as DefaultEventDataSink).IsEnabled = () => SettingsService.GetValue<TestSettings, bool>(x => x.EnableEventRegistering);
+            }));
+            UseModule(new HCTestsModule(new HCTestsModuleOptions() { AssemblyContainingTests = typeof(DevController).Assembly }))
+                .ConfigureGroups((options) => options
+                    .ConfigureGroup(RuntimeTestConstants.Group.AdminStuff, uiOrder: 100)
+                    .ConfigureGroup(RuntimeTestConstants.Group.AlmostTopGroup, uiOrder: 50)
+                    .ConfigureGroup(RuntimeTestConstants.Group.AlmostBottomGroup, uiOrder: -20)
+                    .ConfigureGroup(RuntimeTestConstants.Group.BottomGroup, uiOrder: -50)
+                );
+            UseModule(new HCEventNotificationsModule(new HCEventNotificationsModuleOptions() { EventSink = EventSink }));
+            UseModule(new HCLogViewerModule(new HCLogViewerModuleOptions() { LogSearcherService = CreateLogSearcherService() }));
+            UseModule(new HCDocumentationModule(new HCDocumentationModuleOptions()
+            {
+                EnableDiagramDetails = true,
+                EnableDiagramSandbox = true,
+                SequenceDiagramService = new DefaultSequenceDiagramService(new DefaultSequenceDiagramServiceOptions()
+                {
+                    DefaultSourceAssemblies = new[] { typeof(DevController).Assembly }
+                }),
+                FlowChartsService = new DefaultFlowChartService(new DefaultFlowChartServiceOptions()
+                {
+                    DefaultSourceAssemblies = new[] { typeof(DevController).Assembly }
+                })
+            }));
+            UseModule(new HCDataflowModule<RuntimeTestAccessRole>(new HCDataflowModuleOptions<RuntimeTestAccessRole>() { DataflowService = DataflowService }));
+            UseModule(new HCAuditLogModule(new HCAuditLogModuleOptions() { AuditEventService = _auditEventService }));
+            UseModule(new HCSiteEventsModule(new HCSiteEventsModuleOptions() { SiteEventService = _siteEventService }));
+            UseModule(new HCRequestLogModule(new HCRequestLogModuleOptions() { RequestLogService = RequestLogServiceAccessor.Current }));
+            UseModule(new HCSettingsModule(new HCSettingsModuleOptions() { SettingsService = SettingsService }));
+            //UseModule(new TestModuleB(), "[tst]");
+            //UseModule(new TestModuleA(), "Dev");
 
             if (!_hasInited)
             {
                 InitOnce();
             }
         }
-
-        public ActionResult TestEvent(int v = 1)
-        {
-            object payload = null;
-            switch(v)
-            {
-                case 3:
-                    payload = new
-                    {
-                        Url = Request.RawUrl,
-                        User = CurrentRequestInformation?.UserName,
-                        SettingValue = SettingsService.GetValue<TestSettings, int>((setting) => setting.IntProp),
-                        ExtraB = "BBBB"
-                    };
-                    break;
-                case 2:
-                    payload = new
-                    {
-                        Url = Request.RawUrl,
-                        User = CurrentRequestInformation?.UserName,
-                        SettingValue = SettingsService.GetValue<TestSettings, int>((setting) => setting.IntProp),
-                        ExtraA = "AAAA"
-                    };
-                    break;
-                case 1:
-                default:
-                    payload = new
-                    {
-                        Url = Request.RawUrl,
-                        User = CurrentRequestInformation?.UserName,
-                        SettingValue = SettingsService.GetValue<TestSettings, int>((setting) => setting.IntProp)
-                    };
-                    break;
-            }
-            Services.EventSink.RegisterEvent("pageload", payload);
-            return Content($"Registered variant #{v}");
-        }
-
-        public ActionResult Logout()
-        {
-            ForceLogout = true;
-            return Content("Logged out");
-        }
-
-        public ActionResult Login()
-        {
-            ForceLogout = false;
-            return Content("Logged in");
-        }
-
-        public class TestSettings
-        {
-            [HealthCheckSetting(description: "Some description here")]
-            public string StringProp { get; set; }
-            public bool BoolProp { get; set; } = false;
-            public int IntProp { get; set; } = 15523;
-
-            [HealthCheckSetting(GroupName = "Service X")]
-            public bool EnableX { get; set; }
-
-            [HealthCheckSetting(GroupName = "Service X")]
-            public string ConnectionString{ get; set; }
-
-            [HealthCheckSetting(GroupName = "Service X")]
-            public int Threads { get; set; } = 2;
-
-            [HealthCheckSetting(GroupName = "Service X")]
-            public int NumberOfThings { get; set; } = 321;
-
-            [HealthCheckSetting(GroupName = "Event Notifications")]
-            public bool EnableEventRegistering { get; set; }
-        }
-
-        private ILogSearcherService CreateLogSearcherService()
-            => new FlatFileLogSearcherService(new FlatFileLogSearcherServiceOptions()
-                    .IncludeLogFilesInDirectory(HostingEnvironment.MapPath("~/App_Data/TestLogs/")));
-
-        private ISiteEventService CreateSiteEventService()
-            => new SiteEventService(new FlatFileSiteEventStorage(HostingEnvironment.MapPath("~/App_Data/SiteEventStorage.json"),
-                maxEventAge: TimeSpan.FromDays(5), delayFirstCleanup: false));
-
-        private IAuditEventStorage CreateAuditEventService()
-            => new FlatFileAuditEventStorage(HostingEnvironment.MapPath("~/App_Data/AuditEventStorage.json"),
-                maxEventAge: TimeSpan.FromDays(30), delayFirstCleanup: false);
-
-        private static bool _hasInited = false;
-        private void InitOnce()
-        {
-            _hasInited = true;
-            Task.Run(() => AddEvents(false));
-        }
         #endregion
 
         #region Overrides
+        protected override void ConfigureAccess(HttpRequestBase request, AccessConfig<RuntimeTestAccessRole> config)
+        {
+            /// MODULES //
+            config.GiveRolesAccessToModule(
+                RuntimeTestAccessRole.Guest | RuntimeTestAccessRole.WebAdmins,
+                TestModuleA.TestModuleAAccessOption.DeleteThing | TestModuleA.TestModuleAAccessOption.EditThing
+            );
+
+            config.GiveRolesAccessToModule(RuntimeTestAccessRole.SystemAdmins, TestModuleB.TestModuleBAccessOption.NumberOne);
+
+            // HCDynamicCodeExecutionModule
+            // Guest => View module. Edit local scripts only
+            config.GiveRolesAccessToModule(RuntimeTestAccessRole.Guest, HCDynamicCodeExecutionModule.AccessOption.Nothing);
+            // WebAdmins => only execute directly
+            config.GiveRolesAccessToModule(RuntimeTestAccessRole.WebAdmins, HCDynamicCodeExecutionModule.AccessOption.ExecuteCustomScript);
+            // SomethingElse => only create new ones
+            config.GiveRolesAccessToModule(RuntimeTestAccessRole.SomethingElse, HCDynamicCodeExecutionModule.AccessOption.CreateNewScriptOnServer);
+            // API => load and execute existing scripts
+            config.GiveRolesAccessToModule(RuntimeTestAccessRole.API,
+                HCDynamicCodeExecutionModule.AccessOption.ExecuteSavedScript | HCDynamicCodeExecutionModule.AccessOption.LoadScriptFromServer);
+
+            config.GiveRolesAccessToModuleWithFullAccess<HCTestsModule>(RuntimeTestAccessRole.WebAdmins);
+            config.GiveRolesAccessToModuleWithFullAccess<HCSettingsModule>(RuntimeTestAccessRole.WebAdmins);
+            config.GiveRolesAccessToModuleWithFullAccess<HCRequestLogModule>(RuntimeTestAccessRole.WebAdmins);
+            config.GiveRolesAccessToModuleWithFullAccess<HCSiteEventsModule>(RuntimeTestAccessRole.WebAdmins);
+            config.GiveRolesAccessToModuleWithFullAccess<HCAuditLogModule>(RuntimeTestAccessRole.WebAdmins);
+            config.GiveRolesAccessToModuleWithFullAccess<HCDataflowModule<RuntimeTestAccessRole>>(RuntimeTestAccessRole.WebAdmins);
+            config.GiveRolesAccessToModuleWithFullAccess<HCDocumentationModule>(RuntimeTestAccessRole.WebAdmins);
+            config.GiveRolesAccessToModuleWithFullAccess<HCLogViewerModule>(RuntimeTestAccessRole.WebAdmins);
+            config.GiveRolesAccessToModuleWithFullAccess<HCEventNotificationsModule>(RuntimeTestAccessRole.WebAdmins);
+            config.GiveRolesAccessToModuleWithFullAccess<HCDynamicCodeExecutionModule>(RuntimeTestAccessRole.SystemAdmins);
+            //////////////
+
+            config.ShowFailedModuleLoadStackTrace = new Maybe<RuntimeTestAccessRole>(RuntimeTestAccessRole.WebAdmins);
+            config.PingAccess = new Maybe<RuntimeTestAccessRole>(RuntimeTestAccessRole.API);
+            config.RedirectTargetOnNoAccess = "/no-access";
+        }
+
         public override ActionResult Index()
         {
-            Services.EventSink.RegisterEvent("pageload", new {
+            EventSink.RegisterEvent("pageload", new {
                 Url = Request.RawUrl,
                 User = CurrentRequestInformation?.UserName,
                 SettingValue = SettingsService.GetValue<TestSettings, int>((setting) => setting.IntProp)
@@ -214,7 +194,7 @@ namespace HealthCheck.DevTest.Controllers
                 {
                     Parallel.ForEach(Enumerable.Range(1, count), (i) =>
                     {
-                        Services.EventSink.RegisterEvent("event_parallel_test", new
+                        EventSink.RegisterEvent("event_parallel_test", new
                         {
                             Number = i,
                             TimeStamp = DateTime.Now,
@@ -245,7 +225,7 @@ namespace HealthCheck.DevTest.Controllers
                 watch.Start();
                 for (int i=0;i< eventCount; i++)
                 {
-                    Services.EventSink.RegisterEvent("thing_imported", new
+                    EventSink.RegisterEvent("thing_imported", new
                     {
                         Code = 9999 + i,
                         DisplayName = $"Some item #{(9999 + i)}",
@@ -258,75 +238,40 @@ namespace HealthCheck.DevTest.Controllers
             return base.Index();
         }
 
-        protected override FrontEndOptionsViewModel GetFrontEndOptions()
-            => new FrontEndOptionsViewModel(EndpointBase)
+        protected override HCFrontEndOptions GetFrontEndOptions()
+            => new HCFrontEndOptions(EndpointBase)
             {
                 ApplicationTitle = "Test Monitor",
                 ApplicationTitleLink = "/?sysadmin=x&webadmin=1",
-                PagePriority = new List<HealthCheckPageType>()
+                EditorConfig = new HCFrontEndOptions.EditorWorkerConfig
                 {
-                    HealthCheckPageType.Tests,
-                    HealthCheckPageType.Overview,
-                    HealthCheckPageType.RequestLog,
-                    HealthCheckPageType.EventNotifications,
-                    HealthCheckPageType.Dataflow,
-                    HealthCheckPageType.Documentation,
-                    HealthCheckPageType.LogViewer,
-                    HealthCheckPageType.AuditLog,
-                    HealthCheckPageType.Settings
-                },
-                ApplyCustomColumnRuleByDefault = true,
-                EnableDiagramSandbox = true,
-                EnableDiagramDetails = true
+                    EditorWorkerUrl = $"{EndpointBase}/getscript?name=editor.worker.js",
+                    JsonWorkerUrl = $"{EndpointBase}/getscript?name=json.worker.js"
+                }
             };
 
-        protected override PageOptions GetPageOptions()
-            => new PageOptions()
+        protected override HCPageOptions GetPageOptions()
+            => new HCPageOptions()
             {
+                PageTitle = "Test Monitor",
                 JavaScriptUrls = new List<string> {
                     $"{EndpointBase}/GetVendorScript",
                     $"{EndpointBase}/GetMainScript",
-                },
-                PageTitle = "Test Monitor"
+                }
             };
-
-        protected override void Configure(HttpRequestBase request)
-        {
-            TestRunner.IncludeExceptionStackTraces = CurrentRequestAccessRoles.HasValue && CurrentRequestAccessRoles.Value.HasFlag(RuntimeTestAccessRole.SystemAdmins);
-            AccessOptions.OverviewPageAccess = new Maybe<RuntimeTestAccessRole>(RuntimeTestAccessRole.Guest);
-            AccessOptions.DocumentationPageAccess = new Maybe<RuntimeTestAccessRole>(RuntimeTestAccessRole.WebAdmins);
-            AccessOptions.TestsPageAccess = new Maybe<RuntimeTestAccessRole>(RuntimeTestAccessRole.WebAdmins | RuntimeTestAccessRole.API);
-            AccessOptions.AuditLogAccess = new Maybe<RuntimeTestAccessRole>(RuntimeTestAccessRole.SystemAdmins);
-            AccessOptions.LogViewerPageAccess = new Maybe<RuntimeTestAccessRole>(RuntimeTestAccessRole.SystemAdmins);
-            AccessOptions.InvalidTestsAccess = new Maybe<RuntimeTestAccessRole>(RuntimeTestAccessRole.SystemAdmins);
-            AccessOptions.SiteEventDeveloperDetailsAccess = new Maybe<RuntimeTestAccessRole>(RuntimeTestAccessRole.SystemAdmins);
-            AccessOptions.RequestLogPageAccess = new Maybe<RuntimeTestAccessRole>(RuntimeTestAccessRole.SystemAdmins);
-            AccessOptions.ClearRequestLogAccess = new Maybe<RuntimeTestAccessRole>(RuntimeTestAccessRole.SystemAdmins);
-            AccessOptions.PingAccess = new Maybe<RuntimeTestAccessRole>(RuntimeTestAccessRole.API);
-            AccessOptions.DataflowPageAccess = new Maybe<RuntimeTestAccessRole>(RuntimeTestAccessRole.WebAdmins);
-            AccessOptions.SettingsPageAccess = new Maybe<RuntimeTestAccessRole>(RuntimeTestAccessRole.SystemAdmins);
-            AccessOptions.EventNotificationsPageAccess = new Maybe<RuntimeTestAccessRole>(RuntimeTestAccessRole.SystemAdmins);
-            AccessOptions.EditEventDefinitionsAccess = new Maybe<RuntimeTestAccessRole>(RuntimeTestAccessRole.SystemAdmins);
-
-            AccessOptions.RedirectTargetOnNoAccess = "/no-access";
-        }
-
-        protected override void SetTestSetGroupsOptions(TestSetGroupsOptions options)
-        {
-            options
-                .SetOptionsFor(RuntimeTestConstants.Group.AdminStuff, uiOrder: 100)
-                .SetOptionsFor(RuntimeTestConstants.Group.AlmostTopGroup, uiOrder: 50)
-                .SetOptionsFor(RuntimeTestConstants.Group.AlmostBottomGroup, uiOrder: -20)
-                .SetOptionsFor(RuntimeTestConstants.Group.BottomGroup, uiOrder: -50);
-        }
 
         protected override RequestInformation<RuntimeTestAccessRole> GetRequestInformation(HttpRequestBase request)
         {
-            Services.EventSink.RegisterEvent("GetRequestInfo", new
+            EventSink.RegisterEvent("GetRequestInfo", new
             {
                 Type = this.GetType().Name,
                 Path = Request?.Path
             });
+
+            if (ForcedRole != null)
+            {
+                return new RequestInformation<RuntimeTestAccessRole>(ForcedRole.Value, "forcedId", "Forced role user");
+            }
 
             var roles = RuntimeTestAccessRole.Guest;
             if (request.QueryString["noaccess"] != null)
@@ -366,40 +311,156 @@ namespace HealthCheck.DevTest.Controllers
         #endregion
 
         #region dev
-        [RequestLogInfo(hide: true)]
-        public FileResult GetMainScript()
-        {
-            var filepath = Path.GetFullPath($@"{HostingEnvironment.MapPath("~")}..\HealthCheck.Frontend\dist\healthcheck.js");
-            return new FileStreamResult(new FileStream(filepath, FileMode.Open), "content-disposition");
-        }
-
+        [HideFromRequestLog]
+        public ActionResult GetMainScript() => LoadFile("healthcheck.js");
+        
         //[OutputCache(Duration = 1200, VaryByParam = "none")]
-        [RequestLogInfo(hide: true)]
-        public FileResult GetVendorScript()
+        [HideFromRequestLog]
+        public ActionResult GetVendorScript() => LoadFile("healthcheck.vendor.js");
+
+        [HideFromRequestLog]
+        public ActionResult GetScript([FromUri]string name) => LoadFile(name);
+
+        private ActionResult LoadFile(string filename)
         {
-            var filepath = Path.GetFullPath($@"{HostingEnvironment.MapPath("~")}..\HealthCheck.Frontend\dist\healthcheck.vendor.js");
+            var filepath = Path.GetFullPath($@"{HostingEnvironment.MapPath("~")}..\HealthCheck.Frontend\dist\{filename}");
+            if (!System.IO.File.Exists(filepath)) return Content("");
             return new FileStreamResult(new FileStream(filepath, FileMode.Open), "content-disposition");
         }
 
-        public async Task<ActionResult> RunHealthChecks()
+        public ActionResult TestEvent(int v = 1)
         {
-            var results = await TestRunner.ExecuteTests(TestDiscoverer, 
-                testFilter: (test) => test.Categories.Contains(RuntimeTestConstants.Categories.ScheduledHealthCheck),
-                siteEventService: Services.SiteEventService,
-                auditEventService: Services.AuditEventService
-            );
-            return CreateJsonResult(results.Select(x => new { Test = x.Test?.Name, Result = x.Message, SiteEventTitle = x.SiteEvent?.Title }));
+            object payload = null;
+            payload = v switch
+            {
+                3 => new
+                {
+                    Url = Request.RawUrl,
+                    User = CurrentRequestInformation?.UserName,
+                    SettingValue = SettingsService.GetValue<TestSettings, int>((setting) => setting.IntProp),
+                    ExtraB = "BBBB"
+                },
+                2 => new
+                {
+                    Url = Request.RawUrl,
+                    User = CurrentRequestInformation?.UserName,
+                    SettingValue = SettingsService.GetValue<TestSettings, int>((setting) => setting.IntProp),
+                    ExtraA = "AAAA"
+                },
+                _ => new
+                {
+                    Url = Request.RawUrl,
+                    User = CurrentRequestInformation?.UserName,
+                    SettingValue = SettingsService.GetValue<TestSettings, int>((setting) => setting.IntProp)
+                },
+            };
+            EventSink.RegisterEvent("pageload", payload);
+            return Content($"Registered variant #{v}");
+        }
+        
+        public ActionResult Logout()
+        {
+            ForceLogout = true;
+            return Content("Logged out");
+        }
+
+        public ActionResult Login()
+        {
+            ForceLogout = false;
+            return Content("Logged in");
+        }
+
+        private static RuntimeTestAccessRole? ForcedRole { get; set; }
+        public ActionResult ForceAccessRole(string name = null)
+        {
+            if (name == null)
+            {
+                ForcedRole = null;
+                return Content($"Cleared forced role.");
+            }
+            else
+            {
+                ForcedRole = (RuntimeTestAccessRole) Enum.Parse(typeof(RuntimeTestAccessRole), name);
+                return Content($"Role set to: {ForcedRole}.");
+            }
+        }
+
+        public class TestSettings
+        {
+            [HealthCheckSetting(description: "Some description here")]
+            public string StringProp { get; set; }
+            public bool BoolProp { get; set; } = false;
+            public int IntProp { get; set; } = 15523;
+
+            [HealthCheckSetting(GroupName = "Service X")]
+            public bool EnableX { get; set; }
+
+            [HealthCheckSetting(GroupName = "Service X")]
+            public string ConnectionString { get; set; }
+
+            [HealthCheckSetting(GroupName = "Service X")]
+            public int Threads { get; set; } = 2;
+
+            [HealthCheckSetting(GroupName = "Service X")]
+            public int NumberOfThings { get; set; } = 321;
+
+            [HealthCheckSetting(GroupName = "Event Notifications")]
+            public bool EnableEventRegistering { get; set; }
+        }
+
+        private ILogSearcherService CreateLogSearcherService()
+            => new FlatFileLogSearcherService(new FlatFileLogSearcherServiceOptions()
+                    .IncludeLogFilesInDirectory(HostingEnvironment.MapPath("~/App_Data/TestLogs/")));
+
+        private ISiteEventService CreateSiteEventService()
+            => new SiteEventService(new FlatFileSiteEventStorage(HostingEnvironment.MapPath("~/App_Data/SiteEventStorage.json"),
+                maxEventAge: TimeSpan.FromDays(5), delayFirstCleanup: false));
+
+        private IAuditEventStorage CreateAuditEventService()
+            => new FlatFileAuditEventStorage(HostingEnvironment.MapPath("~/App_Data/AuditEventStorage.json"),
+                maxEventAge: TimeSpan.FromDays(30), delayFirstCleanup: false);
+
+        private static bool _hasInited = false;
+        private void InitOnce()
+        {
+            _hasInited = true;
+            Task.Run(() => AddEvents());
+        }
+
+        private void InitServices()
+        {
+            if (_siteEventService == null)
+            {
+                _siteEventService = CreateSiteEventService();
+                _auditEventService = CreateAuditEventService();
+            }
+
+            DataflowService = new DefaultDataflowService<RuntimeTestAccessRole>(new DefaultDataflowServiceOptions<RuntimeTestAccessRole>()
+            {
+                Streams = new IDataflowStream<RuntimeTestAccessRole>[]
+                {
+                    testStreamA,
+                    testStreamB,
+                    testStreamC,
+                    simpleStream,
+                    memoryStream,
+                    otherStream1,
+                    otherStream2
+                }
+            });
+            EventSink = new DefaultEventDataSink(EventSinkNotificationConfigStorage, EventSinkNotificationDefinitionStorage)
+                .AddNotifier(new WebHookEventNotifier())
+                .AddNotifier(new MyNotifier())
+                .AddNotifier(new SimpleNotifier())
+                .AddPlaceholder("NOW", () => DateTime.Now.ToString())
+                .AddPlaceholder("ServerName", () => Environment.MachineName);
+            (EventSink as DefaultEventDataSink).IsEnabled = () => SettingsService.GetValue<TestSettings, bool>(x => x.EnableEventRegistering);
         }
 
         // New mock data
-        public async Task<ActionResult> AddEvents(bool reInitService = true)
+        public async Task<ActionResult> AddEvents()
         {
-            if (reInitService)
-            {
-                Services.SiteEventService = CreateSiteEventService();
-            }
-
-            if ((await Services.SiteEventService.GetEvents(DateTime.MinValue, DateTime.MaxValue)).Count == 0)
+            if ((await _siteEventService.GetEvents(DateTime.MinValue, DateTime.MaxValue)).Count == 0)
             {
                 for (int i = 0; i < 20; i++)
                 {
@@ -432,7 +493,7 @@ namespace HealthCheck.DevTest.Controllers
         private static readonly Random _rand = new Random();
         public async Task<ActionResult> AddEvent()
         {
-            if (!Enabled || Services.SiteEventService == null) return HttpNotFound();
+            if (!Enabled || _siteEventService == null) return HttpNotFound();
 
             CreateSomeData(out string title, out string description);
             var severity = (_rand.Next(100) < 10) ? SiteEventSeverity.Fatal
@@ -451,7 +512,7 @@ namespace HealthCheck.DevTest.Controllers
             .AddRelatedLink("Page that failed", "https://www.google.com?etc")
             .AddRelatedLink("Error log", "https://www.google.com?q=errorlog");
 
-            await Services.SiteEventService.StoreEvent(ev);
+            await _siteEventService.StoreEvent(ev);
             return CreateJsonResult(ev);
         }
 
