@@ -1,5 +1,5 @@
-﻿using HealthCheck.Core.Modules.Dataflow;
-using HealthCheck.WebUI.Services;
+﻿using HealthCheck.Core.Modules.Dataflow.Abstractions;
+using HealthCheck.Core.Util;
 using System;
 using System.Collections.Generic;
 
@@ -9,7 +9,11 @@ namespace HealthCheck.DevTest._TestImplementation.Dataflow
     {
         public DateTime? InsertionTime { get; set; }
 
-        public string Icon => "<svg xmlns=\"http://www.w3.org/2000/svg\" height=\"24\" viewBox=\"0 0 24 24\" width=\"24\"><path d=\"M9 11.75c-.69 0-1.25.56-1.25 1.25s.56 1.25 1.25 1.25 1.25-.56 1.25-1.25-.56-1.25-1.25-1.25zm6 0c-.69 0-1.25.56-1.25 1.25s.56 1.25 1.25 1.25 1.25-.56 1.25-1.25-.56-1.25-1.25-1.25zM12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm0 18c-4.41 0-8-3.59-8-8 0-.29.02-.58.05-.86 2.36-1.05 4.23-2.98 5.21-5.37C11.07 8.33 14.05 10 17.42 10c.78 0 1.53-.09 2.25-.26.21.71.33 1.47.33 2.26 0 4.41-3.59 8-8 8z\"/><path d=\"M0 0h24v24H0z\" fill=\"none\"/></svg>";
+        public string Icon =>
+            (InsertionTime == null) ? MaterialIcons.AllIcons.Error :
+                (InsertionTime.Value.Ticks % 2 == 0)
+                ? MaterialIcons.Grouped.Navigation.Check
+                : MaterialIcons.Grouped.Navigation.Close;
         public string Code { get; set; }
         public string Name { get; set; }
         public string PreformattedTest => "\tSomething\n\t\there!";
