@@ -205,6 +205,7 @@
 
         <v-dialog v-model="notifierDialogVisible"
             scrollable
+            @keydown.esc="notifierDialogVisible = false"
             v-if="notifiers != null"
             content-class="possible-notifiers-dialog">
             <v-card>
@@ -236,6 +237,7 @@
         </v-dialog>
 
         <v-dialog v-model="deleteDialogVisible"
+            @keydown.esc="deleteDialogVisible = false"
             max-width="290"
             content-class="confirm-dialog">
             <v-card>
@@ -253,6 +255,7 @@
         </v-dialog>
 
         <v-dialog v-model="placeholderDialogVisible"
+            @keydown.esc="placeholderDialogVisible = false"
             scrollable
             content-class="possible-placeholders-dialog">
             <v-card>
@@ -289,7 +292,7 @@ import { Vue, Component, Prop, Watch } from "vue-property-decorator";
 import { EventSinkNotificationConfigFilter, FilterMatchType, EventSinkNotificationConfig, NotifierConfig, Dictionary, IEventNotifier, NotifierConfigOptionsItem, KnownEventDefinition } from "../../models/EventNotifications/EventNotificationModels";
 import SimpleDateTimeComponent from '.././Common/SimpleDateTimeComponent.vue';
 import ConfigFilterComponent from '.././EventNotifications/ConfigFilterComponent.vue';
-import FrontEndOptionsViewModel from "../../models/Page/FrontEndOptionsViewModel";
+import FrontEndOptionsViewModel from "../../models/Common/FrontEndOptionsViewModel";
 import DateUtils from "../../util/DateUtils";
 import IdUtils from "../../util/IdUtils";
 import EventSinkNotificationConfigUtils, { ConfigFilterDescription, ConfigDescription, ConfigActionDescription } from "../../util/EventNotifications/EventSinkNotificationConfigUtils";
@@ -307,8 +310,8 @@ import EventNotificationService from "../../services/EventNotificationService";
 })
 export default class EventNotificationConfigComponent extends Vue {
     @Prop({ required: true })
-    options!: FrontEndOptionsViewModel;
-    
+    moduleId!: string;
+
     @Prop({ required: true })
     config!: EventSinkNotificationConfig;
 
@@ -326,7 +329,7 @@ export default class EventNotificationConfigComponent extends Vue {
 
     // @ts-ignore
     internalConfig: EventSinkNotificationConfig = null;
-    service: EventNotificationService = new EventNotificationService(this.options);
+    service: EventNotificationService = new EventNotificationService(this.globalOptions.InvokeModuleMethodEndpoint, this.globalOptions.InludeQueryStringInApiCalls, this.moduleId);
     ASD!: EventSinkNotificationConfig;
     notifierDialogVisible: boolean = false;
     deleteDialogVisible: boolean = false;
@@ -356,6 +359,10 @@ export default class EventNotificationConfigComponent extends Vue {
     ////////////////
     //  GETTERS  //
     //////////////
+    get globalOptions(): FrontEndOptionsViewModel {
+        return this.$store.state.globalOptions;
+    }
+    
     get allowChanges(): boolean {
         return !this.readonly && !this.serverInteractionInProgress;
     }
