@@ -3,12 +3,19 @@ import DiagramsDataViewModel from "../models/Documentation/DiagramsDataViewModel
 
 export default class DocumentationService extends HCServiceBase
 {
+    public moduleId: string;
+
+    constructor(endpoint: string, inludeQueryString: boolean, moduleId: string)
+    {
+        super(endpoint, inludeQueryString);
+        this.moduleId = moduleId;
+    }
+    
     public GetDiagramsData(
         statusObject: FetchStatus | null = null,
         callbacks: ServiceFetchCallbacks<DiagramsDataViewModel> | null = null
     ) : void
     {
-        let url = this.options.DiagramsDataEndpoint;
-        this.fetchExt<DiagramsDataViewModel>(url, 'GET', null, statusObject, callbacks);
+        this.invokeModuleMethod(this.moduleId, 'GetDiagrams', null, statusObject, callbacks);
     }
 }
