@@ -3,25 +3,31 @@ import { GetEventNotificationConfigsViewModel, EventSinkNotificationConfig } fro
 
 export default class EventNotificationService extends HCServiceBase
 {
+    public moduleId: string;
+
+    constructor(endpoint: string, inludeQueryString: boolean, moduleId: string)
+    {
+        super(endpoint, inludeQueryString);
+        this.moduleId = moduleId;
+    }
+    
     public GetEventNotifications(
         statusObject: FetchStatus | null = null,
         callbacks: ServiceFetchCallbacks<GetEventNotificationConfigsViewModel> | null = null
     ) : void
     {
-        let url = this.options.GetEventNotificationConfigsEndpoint;
-        this.fetchExt<GetEventNotificationConfigsViewModel>(url, 'GET', null, statusObject, callbacks);
+        this.invokeModuleMethod(this.moduleId, 'GetEventNotificationConfigs', null, statusObject, callbacks);
     }
     
     public SetConfigEnabled(config: EventSinkNotificationConfig, enabled: boolean,
         statusObject: FetchStatus | null = null,
         callbacks: ServiceFetchCallbacks<any> | null = null
     ): void {
-        let url = this.options.SetEventNotificationConfigEnabledEndpoint;
         let payload = {
-            configId: config.Id,
-            enabled: enabled
+            ConfigId: config.Id,
+            Enabled: enabled
         };
-        this.fetchExt<any>(url, 'POST', payload, statusObject, callbacks);
+        this.invokeModuleMethod(this.moduleId, 'SetEventNotificationConfigEnabled', payload, statusObject, callbacks);
     }
 
     public SaveConfig(config: EventSinkNotificationConfig,
@@ -29,11 +35,7 @@ export default class EventNotificationService extends HCServiceBase
         callbacks: ServiceFetchCallbacks<EventSinkNotificationConfig> | null = null
     ): void
     {
-        let url = this.options.SaveEventNotificationConfigEndpoint;
-        let payload = {
-            config: config
-        };
-        this.fetchExt<EventSinkNotificationConfig>(url, 'POST', payload, statusObject, callbacks);
+        this.invokeModuleMethod(this.moduleId, 'SaveEventNotificationConfig', config, statusObject, callbacks);
     }
 
     public DeleteConfig(configId: string,
@@ -41,11 +43,7 @@ export default class EventNotificationService extends HCServiceBase
         callbacks: ServiceFetchCallbacks<EventSinkNotificationConfig> | null = null
     ): void
     {
-        let url = this.options.DeleteEventNotificationConfigEndpoint;
-        let payload = {
-            configId: configId
-        };
-        this.fetchExt<EventSinkNotificationConfig>(url, 'POST', payload, statusObject, callbacks);
+        this.invokeModuleMethod(this.moduleId, 'DeleteEventNotificationConfig', configId, statusObject, callbacks);
     }
 
     public DeleteDefintion(eventId: string,
@@ -53,11 +51,7 @@ export default class EventNotificationService extends HCServiceBase
         callbacks: ServiceFetchCallbacks<boolean> | null = null
     ): void
     {
-        let url = this.options.DeleteEventDefinitionEndpoint;
-        let payload = {
-            eventId: eventId
-        };
-        this.fetchExt<boolean>(url, 'POST', payload, statusObject, callbacks);
+        this.invokeModuleMethod(this.moduleId, 'DeleteEventDefinition', eventId, statusObject, callbacks);
     }
 
     public DeleteDefintions(
@@ -65,7 +59,6 @@ export default class EventNotificationService extends HCServiceBase
         callbacks: ServiceFetchCallbacks<boolean> | null = null
     ): void
     {
-        let url = this.options.DeleteEventDefinitionsEndpoint;
-        this.fetchExt<boolean>(url, 'POST', null, statusObject, callbacks);
+        this.invokeModuleMethod(this.moduleId, 'DeleteEventDefinitions', null, statusObject, callbacks);
     }
 }
