@@ -4,27 +4,32 @@ import LogSearchResult from "../models/LogViewer/LogSearchResult";
 
 export default class LogService extends HCServiceBase
 {
+    public moduleId: string;
+
+    constructor(endpoint: string, inludeQueryString: boolean, moduleId: string)
+    {
+        super(endpoint, inludeQueryString);
+        this.moduleId = moduleId;
+    }
+    
     public Search(filter: Partial<LogSearchFilter>,
         statusObject: FetchStatus | null = null,
         callbacks: ServiceFetchCallbacks<LogSearchResult> | null = null
     ): void {
-        let url = this.options.GetLogSearchResultsEndpoint;
-        this.fetchExt<any>(url, 'POST', filter, statusObject, callbacks);
+        this.invokeModuleMethod(this.moduleId, 'SearchLogs', filter, statusObject, callbacks);
     }
     
     public CancelSearch(searchId: string,
         statusObject: FetchStatus | null = null,
         callbacks: ServiceFetchCallbacks<boolean> | null = null
     ): void {
-        let url = this.options.CancelLogSearchEndpoint;
-        this.fetchExt<boolean>(url, 'POST', { searchId: searchId }, statusObject, callbacks);
+        this.invokeModuleMethod(this.moduleId, 'CancelLogSearch', searchId, statusObject, callbacks);
     }
     
     public CancelAllSearches(
         statusObject: FetchStatus | null = null,
         callbacks: ServiceFetchCallbacks<number> | null = null
     ): void {
-        let url = this.options.CancelAllLogSearchesEndpoint;
-        this.fetchExt<number>(url, 'POST', null, statusObject, callbacks);
+        this.invokeModuleMethod(this.moduleId, 'CancelAllLogSearches', null, statusObject, callbacks);
     }
 }
