@@ -1,5 +1,5 @@
-﻿using HealthCheck.Core.Abstractions;
-using HealthCheck.Core.Entities;
+﻿using HealthCheck.Core.Modules.AuditLog.Abstractions;
+using HealthCheck.Core.Modules.AuditLog.Models;
 using HealthCheck.Core.Util;
 using Newtonsoft.Json;
 using System;
@@ -59,7 +59,7 @@ namespace HealthCheck.WebUI.Services
         public Task<List<AuditEvent>> GetEvents(DateTime from, DateTime to)
         {
             var items = Store.GetEnumerable()
-                .Where(x => x.Timestamp >= from && x.Timestamp <= to)
+                .Where(x => x.Timestamp.ToUniversalTime() >= from && x.Timestamp.ToUniversalTime() <= to)
                 .ToList();
 
             return Task.FromResult(items);
